api-client: move base api and session to internal subfolder
